requirements.txt
flask==3.1.1
langchain-pinecone==0.2.8
langchain
langchain-openai
python-dotenv
langchain-groq
google-generativeai
langchain-google-genai
langchain-core
langchain-community
pypdf
tiktoken
chromadb
faiss-cpu
langchain-huggingface
sentence-transformers
transformers
torch
langchain_core